Implements LineReader, which reads lines from a file expected to be appended continuously.
This is not useful by itself, but I figured to get some feedback before making more progress.

The key features of LineReader are:

It reports lines with positions so that reading can be resumed later. possibly in another process.
It deals with EOF in the expectation of files being externally appended
to.